backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
openai==1.86.0
python-multipart==0.0.20
python-dotenv==1.1.0
pydantic==2.11.7
requests==2.32.4
numpy==2.3.0
faiss-cpu==1.11.0
langchain==0.3.25
langchain-openai==0.3.23
PyPDF2==3.0.1
python-docx==1.1.2
pandas==2.2.3
openpyxl==3.1.5
psutil==6.1.0
